refactor: CartManager 제거 및 CartViewModel에 해당 내용 병합
